pyproject.toml
[project]
name = "biosym"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "absl-py>=2.3.0",
    "black[jupyter]>=25.1.0",
    "cloudpickle>=3.1.1",
    "flatbuffers>=25.2.10",
    "flax>=0.8.5",
    "jax>=0.4.30",
    "numpy>=2.0.2",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "ruff>=0.12.1",
    "sympy>=1.14.0",
]
